requirements.txt
appdirs==1.4.3
Django==1.11.3
django-crispy-forms==1.6.1
django-mptt==0.8.7
django-polymorphic==1.1
packaging==16.8
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
Unidecode==0.4.20
